% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statSummary.R
\name{statSummary}
\alias{statSummary}
\title{Generate numeric summaries}
\usage{
statSummary(x, na.rm = TRUE)
}
\arguments{
\item{x}{The numeric data}

\item{na.rm}{Should missing values be removed?}
}
\description{
Generates numeric summaries, removing missing values
}
\author{
Aimee Gott
}

